requirements.txt
Mako==0.7.3
MarkupSafe==1.0
ansible>=2.7.12
certifi==2018.4.16
cffi==1.8.2
click==6.7
configparser==3.5.0
eventlet==0.24.1
Flask-Cors==3.0.4
flask-restful-swagger-2==0.35
Flask-SocketIO==4.3.2
flask-swagger-ui==3.0.12
flask==1.0.2
flask_restful==0.3.6
futures==3.0.5
gevent-websocket==0.10.1
gevent==1.2.2
gunicorn==19.8.1
hurry.filesize==0.9
mixpanel==4.3.2
netifaces==0.10.4
paho.mqtt==1.4.0
pyasn1==0.1.8
python-dateutil==2.4.2
pytz==2012d
requests[security]==2.22.0
sh==1.08
six==1.11.0
uptime==2.0.2
urllib3>=1.24.2
werkzeug>=0.15.3
youtube_dl>=2017.7.2
